StubProgram/requirements.txt
pycryptodome==3.9.9
pyffx==0.3.0
websockets==8.1
